requirements.txt
attrs==20.3.0
Authlib==0.15.2
bastion-key-swagger-client==1.0.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
charset-normalizer==2.0.10
click==7.1.2
colorama==0.4.4
cryptography==3.3.2
docutils==0.17.1
Flask==1.1.2
idna==2.10
importlib-metadata==3.7.3
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
keyring==21.5.0
MarkupSafe==1.1.1
pluggy==0.13.1
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.9.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.2
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
toml==0.10.2
tqdm==4.54.1
urllib3==1.26.5
webencodings==0.5.1
Werkzeug==1.0.1
zipp==3.4.0
python-dotenv==0.19.2

